# docs/api/openapi-v1.designed.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "QueryRequest": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "domain": {
                  "const": "document"
                }
              },
              "required": [
                "domain"
              ]
            },
            "then": {
              "properties": {
                "document_ids": {
                  "type": "array"
                }
              },
              "required": [
                "document_ids"
              ]
            }
          },
          {
            "if": {
              "anyOf": [
                {
                  "not": {
                    "required": [
                      "domain"
                    ]
                  }
                },
                {
                  "properties": {
                    "domain": {
                      "const": "default"
                    }
                  }
                }
              ]
            },
            "then": {
              "not": {
                "required": [
                  "document_ids"
                ]
              }
            }
          }
        ],
        "properties": {
          "answer_language": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "vi"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Answer Language"
          },
          "corpus_languages": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "en",
                    "vi"
                  ],
                  "type": "string"
                },
                "maxItems": 2,
                "minItems": 1,
                "type": "array",
                "uniqueItems": true
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Corpus Languages"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 50,
                "minItems": 1,
                "type": "array",
                "uniqueItems": true
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Document Ids"
          },
          "domain": {
            "default": "default",
            "enum": [
              "default",
              "document",
              "multilingual"
            ],
            "title": "Domain",
            "type": "string"
          },
          "history": {
            "description": "Untrusted conversation, never factual evidence. Above-budget history is accepted; runtime must truncate and report counts, not silently reject or retain all.",
            "items": {
              "$ref": "#/components/schemas/HistoryMessage"
            },
            "title": "History",
            "type": "array"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "question"
        ],
        "title": "QueryRequest",
        "type": "object",
        "x-runtime-enforcement": [
          "authenticated principal/current session/ready versions/subset ownership",
          "history truncation with metadata at 20 messages/8000 tokenizer tokens",
          "context 8000/output 1024 tokenizer tokens; scope revalidation"
        ]
      },
...
      "SSEEvent": {
        "discriminator": {
          "mapping": {
            "answer_delta": "#/components/schemas/AnswerDeltaEvent",
            "done": "#/components/schemas/DoneEvent",
            "error": "#/components/schemas/ErrorEvent",
            "evidence": "#/components/schemas/EvidenceEvent",
            "meta": "#/components/schemas/MetaEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/MetaEvent"
          },
          {
            "$ref": "#/components/schemas/EvidenceEvent"
          },
          {
            "$ref": "#/components/schemas/AnswerDeltaEvent"
          },
          {
            "$ref": "#/components/schemas/DoneEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "AppServiceKey": {
        "in": "header",
        "name": "X-RAG-Service-Key",
        "type": "apiKey"
      },
      "UserJWT": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API v1 contract inventory. Public routes mounted in T26; business services require trusted configuration and authentication.",
    "title": "RAG Core API v1 — designed contract",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/query": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "query",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T24/T26",
        "x-served": true
      }
    },
    "/v1/query/stream": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "query_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SSEEvent"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "query stream",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T25/T26",
        "x-served": true,
        "x-sse-contract": {
          "deltas": "provisional",
          "done": "validated final QueryResponse",
          "early-error": "meta -> error before evidence is allowed",
          "event-ids": "increasing per request",
          "heartbeat": "SSE comment; no event ID",
          "order": "meta -> evidence -> answer_delta* -> done|error",
          "replay": false,
          "runtime": "IMPLEMENTED; scope checks before evidence/delta/done; cancel upstream",
          "wire": "SSE UTF-8 frames: id/event/data lines, blank-line delimiter; data line contains JSON. SSEEvent schema describes a parsed event, not a JSON HTTP response body."
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "create session",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T10",
        "x-served": true
      }
    },
...
    "/v1/sessions/{session_id}/documents": {
      "get": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "list_documents",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 50,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "list documents",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T12",
        "x-served": true
      },
      "post": {
        "description": "Mounted in T26; trusted runtime configuration is required. Authenticated app+subject and active session links/ready versions constrain access. Schemas do not grant ownership or readiness.",
        "operationId": "register_document",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Scoped app+owner+session+request hash; mismatch 409",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRegisterResponse"
                }
              }
            },
            "description": "Served response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid or missing credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Insufficient role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not in current authorized scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Scope/readiness/idempotency conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Deleted session for its owner"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or bounded queue full",
            "headers": {
              "Retry-After": {
                "description": "Backoff delay in seconds",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Provider error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dependency unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Timeout"
          }
        },
        "security": [
          {
            "AppServiceKey": [],
            "UserJWT": []
          }
        ],
        "summary": "register document",
        "tags": [
          "public"
        ],
        "x-implementation-status": "IMPLEMENTED",
        "x-implementation-task": "T12",
        "x-served": true
      }
    },
...
  },
...
}